+++ src/api_v3/signup/render_form_signup.js
